[DP-938] fix
